requirements.txt
python-telegram-bot
qrcode
pillow